Changed UserRouter into AccountRouter, moved methods
